Abstract functionality used to build @with directives
