+++ src/Schema/Directives/WithCountDirective.php
@@ -2,19 +2,12 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Closure;
-use GraphQL\Deferred;
-use GraphQL\Type\Definition\ResolveInfo;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
-use Nuwave\Lighthouse\Execution\DataLoader\BatchLoader;
 use Nuwave\Lighthouse\Execution\DataLoader\RelationCountBatchLoader;
-use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
-class WithCountDirective extends RelationDirective implements FieldMiddleware, DefinedDirective
+class WithCountDirective extends WithRelationDirective implements FieldMiddleware, DefinedDirective
 {
     public static function definition(): string
     {
@@ -38,53 +31,15 @@ SDL;
     }
 
     /**
-     * Eager load a count of a relation on the parent instance.
+     * The name of the batch loader to use.
      */
-    public function handleField(FieldValue $fieldValue, Closure $next): FieldValue
+    public function batchLoaderName(): string
     {
-        $resolver = $fieldValue->getResolver();
-
-        return $next(
-            $fieldValue->setResolver(
-                function (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($resolver): Deferred {
-                    return new Deferred(function () use ($resolver, $parent, $args, $context, $resolveInfo) {
-                        return $this->loader($resolveInfo)
-                            ->load($parent->getKey(), ['parent' => $parent])
-                            ->then(
-                                function () use ($resolver, $parent, $args, $context, $resolveInfo) {
-                                    return $resolver($parent, $args, $context, $resolveInfo);
-                                }
-                            );
-                    });
-                }
-            )
-        );
+        return RelationCountBatchLoader::class;
     }
 
     /**
-     * Create an instance of RelationCountBatchLoader loader to apply counts.
-     */
-    protected function loader(ResolveInfo $resolveInfo): RelationCountBatchLoader
-    {
-        return BatchLoader::instance( // @phpstan-ignore-line TODO remove when updating graphql-php
-            RelationCountBatchLoader::class,
-            $resolveInfo->path,
-            [
-                'relationName' => $this->relationName(),
-                'decorateBuilder' => function ($query) use ($resolveInfo) {
-                    $resolveInfo
-                        ->argumentSet
-                        ->enhanceBuilder(
-                            $query,
-                            $this->directiveArgValue('scopes', [])
-                        );
-                },
-            ]
-        );
-    }
-
-    /**
-     * The the name of the relation to be counted.
+     * The the name of the relation to be loaded.
      */
     protected function relationName(): string
     {
