+++ src/Schema/Directives/WithDirective.php
@@ -2,18 +2,12 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Closure;
-use GraphQL\Deferred;
-use GraphQL\Type\Definition\ResolveInfo;
-use Illuminate\Database\Eloquent\Model;
-use Nuwave\Lighthouse\Execution\DataLoader\BatchLoader;
 use Nuwave\Lighthouse\Execution\DataLoader\RelationBatchLoader;
-use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
-class WithDirective extends RelationDirective implements FieldMiddleware, DefinedDirective
+class WithDirective extends WithRelationDirective implements FieldMiddleware, DefinedDirective
 {
     public static function definition(): string
     {
@@ -37,45 +31,10 @@ SDL;
     }
 
     /**
-     * Eager load a relation on the parent instance.
+     * The name of the batch loader to use.
      */
-    public function handleField(FieldValue $fieldValue, Closure $next): FieldValue
+    public function batchLoaderName(): string
     {
-        $resolver = $fieldValue->getResolver();
-
-        return $next(
-            $fieldValue->setResolver(
-                function (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($resolver): Deferred {
-                    $loader = BatchLoader::instance( // @phpstan-ignore-line TODO remove when updating graphql-php
-                        RelationBatchLoader::class,
-                        $resolveInfo->path,
-                        [
-                            'relationName' => $this->directiveArgValue('relation', $this->nodeName()),
-                            'decorateBuilder' => function ($query) use ($resolveInfo) {
-                                $resolveInfo
-                                    ->argumentSet
-                                    ->enhanceBuilder(
-                                        $query,
-                                        $this->directiveArgValue('scopes', [])
-                                    );
-                            },
-                        ]
-                    );
-
-                    return new Deferred(function () use ($loader, $resolver, $parent, $args, $context, $resolveInfo) {
-                        return $loader
-                            ->load(
-                                $parent->getKey(),
-                                ['parent' => $parent]
-                            )
-                            ->then(
-                                function () use ($resolver, $parent, $args, $context, $resolveInfo) {
-                                    return $resolver($parent, $args, $context, $resolveInfo);
-                                }
-                            );
-                    });
-                }
-            )
-        );
+        return RelationBatchLoader::class;
     }
 }
